+++ tests/org/cicirello/search/problems/scheduling/HeuristicTests.java
@@ -628,6 +628,38 @@ public class HeuristicTests {
 	}
 	
 	@Test
+	public void testSPTSetupAdjustedPrecomputed() {
+		double e = ShortestProcessingPlusSetupTimePrecompute.MIN_H;
+		int highP = (int)Math.ceil(1 / e)*2;
+		int[] w =    { 1, 1, 1, 1, 0, 0, 0, 0, 2, 2, 2, 2, 1};
+		int[] p =    { 1, 2, 4, 8, 1, 2, 4, 8, 1, 2, 4, 8, highP};
+		double[] expected = { 1, 0.5, 0.25, 0.125, 1, 0.5, 0.25, 0.125, 1, 0.5, 0.25, 0.125, e};
+		PartialPermutation partial = new PartialPermutation(expected.length);
+		FakeProblemWeightsPTime problem = new FakeProblemWeightsPTime(w, p);
+		ShortestProcessingPlusSetupTimePrecompute h = new ShortestProcessingPlusSetupTimePrecompute(problem);
+		for (int j = 0; j < expected.length; j++) {
+			assertEquals(expected[j], h.h(partial, j, null), 1E-10);
+		}
+		partial.extend(p.length-1);
+		for (int j = 0; j < expected.length-1; j++) {
+			assertEquals(expected[j], h.h(partial, j, null), 1E-10);
+		}
+		
+		int[] ps =    { 0, 1, 3, 7, 0, 1, 3, 7, 0, 1, 3, 7, highP - 1};
+		problem = new FakeProblemWeightsPTime(w, ps, 0, 1);
+		h = new ShortestProcessingPlusSetupTimePrecompute(problem);
+		partial = new PartialPermutation(expected.length);
+		for (int j = 0; j < expected.length; j++) {
+			assertEquals(expected[j], h.h(partial, j, null), 1E-10);
+		}
+		partial.extend(4);
+		double[] expected2 = {1.0/8, 1.0/10, 1.0/13, 1.0/18};
+		for (int j = 0; j < expected2.length; j++) {
+			assertEquals(expected2[j], h.h(partial, j, null), 1E-10);
+		}
+	}
+	
+	@Test
 	public void testWSPT2() {
 		double e = WeightedShortestProcessingTimeLateOnly.MIN_H;
 		int highP = (int)Math.ceil(1 / e)*2;